Shrink size of attributes processor from early prototype.
