+++ sdk/metrics/src/main/java/io/opentelemetry/sdk/metrics/internal/view/AttributesProcessor.java
@@ -10,8 +10,12 @@ import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.context.Context;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
+import java.util.function.BiFunction;
 import java.util.function.Predicate;
+import java.util.function.UnaryOperator;
 import javax.annotation.concurrent.Immutable;
 
 /**
@@ -21,35 +25,34 @@ import javax.annotation.concurrent.Immutable;
  * <p>An AttributesProcessor is used to define the actual set of attributes that will be used in a
  * Metric vs. the inbound set of attributes from a measurement.
  */
-@FunctionalInterface
 @Immutable
-public interface AttributesProcessor {
+public abstract class AttributesProcessor {
+  private AttributesProcessor() {}
 
   /**
    * Manipulates a set of attributes, returning the desired set.
    *
-   * @param incoming Attributes assocaited with an incoming measurement.
+   * @param incoming Attributes associated with an incoming measurement.
    * @param context The context associated with the measurement.
    */
-  Attributes process(Attributes incoming, Context context);
+  public abstract Attributes process(Attributes incoming, Context context);
 
   /**
-   * If true, this ensures the `Context` argument of the attributes processor is always acurate.
+   * If true, this ensures the `Context` argument of the attributes processor is always accurate.
    * This will prevents bound instruments from pre-locking their metric-attributes and defer until
    * context is available.
    */
-  default boolean usesContext() {
-    return true;
-  }
+  public abstract boolean usesContext();
 
   /** Joins this attribute processor with another that operates after this one. */
-  default AttributesProcessor then(AttributesProcessor other) {
+  public AttributesProcessor then(AttributesProcessor other) {
     if (other == NOOP) {
       return this;
     }
     if (this == NOOP) {
       return other;
     }
+
     if (other instanceof JoinedAttribtuesProcessor) {
       return ((JoinedAttribtuesProcessor) other).prepend(this);
     }
@@ -68,24 +71,17 @@ public interface AttributesProcessor {
    */
   @SuppressWarnings("unchecked")
   public static AttributesProcessor filterByKeyName(Predicate<String> nameFilter) {
-    return new SimpleAttributesProcessor() {
-      @Override
-      protected Attributes process(Attributes incoming) {
-        AttributesBuilder result = Attributes.builder();
-        incoming.forEach(
-            (k, v) -> {
-              if (nameFilter.test(k.getKey())) {
-                result.put((AttributeKey<Object>) k, v);
-              }
-            });
-        return result.build();
-      }
-
-      @Override
-      public String toString() {
-        return "KeyNameFilterProcessor";
-      }
-    };
+    return simple(
+        incoming -> {
+          AttributesBuilder result = Attributes.builder();
+          incoming.forEach(
+              (k, v) -> {
+                if (nameFilter.test(k.getKey())) {
+                  result.put((AttributeKey<Object>) k, v);
+                }
+              });
+          return result.build();
+        });
   }
 
   /**
@@ -96,27 +92,19 @@ public interface AttributesProcessor {
    * @param nameFilter a filter for which baggage keys to select.
    */
   public static AttributesProcessor appendBaggageByKeyName(Predicate<String> nameFilter) {
-    return new BaggageAttributesProcessor() {
-
-      @Override
-      protected Attributes process(Attributes incoming, Baggage baggage) {
-        AttributesBuilder result = Attributes.builder();
-        baggage.forEach(
-            (k, v) -> {
-              if (nameFilter.test(k)) {
-                result.put(k, v.getValue());
-              }
-            });
-        // Override any baggage keys with existing keys.
-        result.putAll(incoming);
-        return result.build();
-      }
-
-      @Override
-      public String toString() {
-        return "BaggageAppendProcessor";
-      }
-    };
+    return onBaggage(
+        (incoming, baggage) -> {
+          AttributesBuilder result = Attributes.builder();
+          baggage.forEach(
+              (k, v) -> {
+                if (nameFilter.test(k)) {
+                  result.put(k, v.getValue());
+                }
+              });
+          // Override any baggage keys with existing keys.
+          result.putAll(incoming);
+          return result.build();
+        });
   }
 
   /**
@@ -127,35 +115,84 @@ public interface AttributesProcessor {
    * @param attributes Attributes to append to measurements.
    */
   public static AttributesProcessor append(Attributes attributes) {
-    return new SimpleAttributesProcessor() {
+    return simple(incoming -> attributes.toBuilder().putAll(incoming).build());
+  }
+
+  /** Creates a simple attributes processor with no access to context. */
+  static AttributesProcessor simple(UnaryOperator<Attributes> processor) {
+    return new AttributesProcessor() {
 
       @Override
-      protected Attributes process(Attributes incoming) {
-        return attributes.toBuilder().putAll(incoming).build();
+      public Attributes process(Attributes incoming, Context context) {
+        return processor.apply(incoming);
       }
 
       @Override
-      public String toString() {
-        return "AppendAttributesProcessor";
+      public boolean usesContext() {
+        return false;
       }
     };
   }
 
-  static final AttributesProcessor NOOP =
-      new AttributesProcessor() {
-        @Override
-        public Attributes process(Attributes incoming, Context context) {
-          return incoming;
-        }
+  /** Creates an Attributes processor that has access to baggage. */
+  static AttributesProcessor onBaggage(BiFunction<Attributes, Baggage, Attributes> processor) {
+    return new AttributesProcessor() {
+      @Override
+      public Attributes process(Attributes incoming, Context context) {
+        return processor.apply(incoming, Baggage.fromContext(context));
+      }
 
-        @Override
-        public boolean usesContext() {
-          return false;
-        }
+      @Override
+      public boolean usesContext() {
+        return true;
+      }
+    };
+  }
 
-        @Override
-        public String toString() {
-          return "NoopAttributesProcessor";
-        }
-      };
+  static final AttributesProcessor NOOP = simple(incoming -> incoming);
+
+  /** A {@link AttributesProcessor} that runs a sequence of processors. */
+  @Immutable
+  static final class JoinedAttribtuesProcessor extends AttributesProcessor {
+    private final Collection<AttributesProcessor> processors;
+    private final boolean usesContextCache;
+
+    JoinedAttribtuesProcessor(Collection<AttributesProcessor> processors) {
+      this.processors = processors;
+      this.usesContextCache =
+          processors.stream().map(AttributesProcessor::usesContext).reduce(false, (l, r) -> l || r);
+    }
+
+    @Override
+    public Attributes process(Attributes incoming, Context context) {
+      Attributes result = incoming;
+      for (AttributesProcessor processor : processors) {
+        result = processor.process(result, context);
+      }
+      return result;
+    }
+
+    @Override
+    public boolean usesContext() {
+      return usesContextCache;
+    }
+
+    @Override
+    public AttributesProcessor then(AttributesProcessor other) {
+      ArrayList<AttributesProcessor> newList = new ArrayList<>(processors);
+      if (other instanceof JoinedAttribtuesProcessor) {
+        newList.addAll(((JoinedAttribtuesProcessor) other).processors);
+      } else {
+        newList.add(other);
+      }
+      return new JoinedAttribtuesProcessor(newList);
+    }
+
+    AttributesProcessor prepend(AttributesProcessor other) {
+      ArrayList<AttributesProcessor> newList = new ArrayList<>(processors.size() + 1);
+      newList.add(other);
+      newList.addAll(processors);
+      return new JoinedAttribtuesProcessor(newList);
+    }
+  }
 }
